move suit, trumpf and deck to rules
