+++ README.md
@@ -98,6 +98,7 @@ which provides some basic functionality like store your cards.
 
 To get more familiar with this concept, let's have a look at the already mentioned 
 Random Player.
+
 ```python
 import random
 
@@ -105,25 +106,25 @@ from schieber.card import Card
 from schieber.game import GameState
 from schieber.player.base_player import BasePlayer
 from schieber.rules.trumpf_rules import trumpf_allowed
-from schieber.trumpf import Trumpf
+from schieber.rules.trumpf import Trumpf
 
 
 class RandomPlayer(BasePlayer):
-    def choose_trumpf(self, state: GameState) -> 'Trumpf':
-        return select_random_trumpf(state.geschoben)
+   def choose_trumpf(self, state: GameState) -> 'Trumpf':
+      return select_random_trumpf(state.geschoben)
 
-    def choose_card(self, state: GameState) -> 'Card':
-        cards = self.allowed_cards(state=state)
-        random.shuffle(cards)
-        return cards[0]
+   def choose_card(self, state: GameState) -> 'Card':
+      cards = self.allowed_cards(state=state)
+      random.shuffle(cards)
+      return cards[0]
 
 
 def select_random_trumpf(geschoben: bool):
-    choices = list(Trumpf)
-    random.shuffle(choices)
-    for choice in choices:
-        if trumpf_allowed(chosen_trumpf=choice, geschoben=geschoben):
-            return choice
+   choices = list(Trumpf)
+   random.shuffle(choices)
+   for choice in choices:
+      if trumpf_allowed(chosen_trumpf=choice, geschoben=geschoben):
+         return choice
 
 ```
 What's going on here?
