+++ schieber/player/base_player.py
@@ -2,7 +2,7 @@ import inspect
 
 from schieber.card import from_string_to_card, Card
 from schieber.game import GameState
-from schieber.trumpf import Trumpf
+from schieber.rules.trumpf import Trumpf
 from schieber.rules.stich_rules import allowed_cards
