+++ schieber/player/challenge_player/strategy/card_counter.py
@@ -1,7 +1,6 @@
 from math import floor
 
-from schieber.card import from_string_to_card
-from schieber.deck import Deck
+from schieber.rules.deck import Deck
 from schieber.player.challenge_player.strategy.flags.doesnt_habe_card_flag import DoesntHaveCardFlag
 from schieber.player.challenge_player.strategy.flags.falied_to_serve_suit_flag import FailedToServeSuitFlag
 from schieber.player.challenge_player.strategy.flags.previously_had_stich_flag import PreviouslyHadStichFlag
@@ -11,7 +10,7 @@ from schieber.player.challenge_player.strategy.mode.top_down_mode import *
 from schieber.player.challenge_player.strategy.mode.trumpf_color_mode import *
 from schieber.rules.stich_rules import stich_rules
 from schieber.stich import PlayedCard
-from schieber.suit import Suit
+from schieber.rules.suit import Suit
 
 
 class CardCounter:
