+++ schieber/player/challenge_player/strategy/mode/top_down_mode.py
@@ -1,7 +1,7 @@
 from schieber.game import GameState
 from schieber.player.challenge_player.strategy.mode.uncolored_trumpf import UncoloredTrumpf
 from schieber.helpers.game_helper import *
-from schieber.trumpf import Trumpf
+from schieber.rules.trumpf import Trumpf
 from schieber.card import from_string_to_card
