+++ schieber/player/random_player.py
@@ -4,7 +4,7 @@ from schieber.card import Card
 from schieber.game import GameState
 from schieber.player.base_player import BasePlayer
 from schieber.rules.trumpf_rules import trumpf_allowed
-from schieber.trumpf import Trumpf
+from schieber.rules.trumpf import Trumpf
 
 
 class RandomPlayer(BasePlayer):
