+++ schieber/rules/deck.py
@@ -1,4 +1,4 @@
-from schieber.suit import Suit
+from schieber.rules.suit import Suit
 from schieber.card import Card
