+++ schieber/rules/suit.py
@@ -1,5 +1,6 @@
 from enum import Enum
 
+
 class Suit(Enum):
     ROSE = 0
     SCHELLE = 1
