+++ tests/player/challenge_player/strategy/mode/test_trumpf_color_mode.py
@@ -1,7 +1,7 @@
 import pytest
 
 from schieber.card import Card
-from schieber.suit import Suit
+from schieber.rules.suit import Suit
 from schieber.player.challenge_player.strategy.mode.trumpf_color_mode import TrumpfColorMode
