+++ tests/player/challenge_player/strategy/test_jass_strategy.py
@@ -2,8 +2,8 @@ import pytest
 
 from schieber.card import Card
 from schieber.game import Game
-from schieber.suit import Suit
-from schieber.trumpf import Trumpf
+from schieber.rules.suit import Suit
+from schieber.rules.trumpf import Trumpf
 from schieber.player.challenge_player.strategy.jass_strategy import JassStrategy
 from schieber.player.challenge_player.challenge_player import ChallengePlayer
