+++ tests/player/greedy_player/test_example_trumpf.py
@@ -1,10 +1,9 @@
 import pytest
 
 from schieber.card import Card
-from schieber.game import Game
 from schieber.player.greedy_player.trumpf_decision import choose_trumpf, TrumpfType
-from schieber.suit import Suit
-from schieber.trumpf import Trumpf
+from schieber.rules.suit import Suit
+from schieber.rules.trumpf import Trumpf
 
 
 @pytest.mark.parametrize("cards, trumpf, trumpf_type", [
