+++ tests/rules/test_stich_rules.py
@@ -2,11 +2,11 @@ import pytest
 
 from schieber.rules.stich_rules import stich_rules, card_allowed, allowed_cards, is_trumpf_under, does_under_trumpf, \
     is_chosen_card_best_trumpf
-from schieber.trumpf import Trumpf
+from schieber.rules.trumpf import Trumpf
 from schieber.card import Card
 from schieber.player.random_player import RandomPlayer
 from schieber.stich import PlayedCard
-from schieber.suit import Suit
+from schieber.rules.suit import Suit
 
 
 @pytest.fixture(scope="module", autouse=True)
